+++ pint_xarray/tests/test_accessors.py
@@ -65,10 +65,10 @@ def example_quantity_da():
 class TestQuantifyDataArray:
     def test_attach_units_from_str(self, example_unitless_da):
         orig = example_unitless_da
-        result = orig.pint.quantify("m")
+        result = orig.pint.quantify("s")
         assert_array_equal(result.data.magnitude, orig.data)
         # TODO better comparisons for when you can't access the unit_registry?
-        assert str(result.data.units) == "meter"
+        assert str(result.data.units) == "second"
 
     def test_attach_units_given_registry(self, example_unitless_da):
         orig = example_unitless_da
@@ -86,12 +86,29 @@ class TestQuantifyDataArray:
         remaining_attrs = conversion.extract_unit_attributes(result)
         assert {k: v for k, v in remaining_attrs.items() if v is not None} == {}
 
+    def test_attach_units_from_str_attr_no_unit(self, example_unitless_da):
+        orig = example_unitless_da
+        orig.attrs["units"] = "none"
+        result = orig.pint.quantify("m")
+        assert_array_equal(result.data.magnitude, orig.data)
+        assert str(result.data.units) == "meter"
+
     def test_attach_units_given_unit_objs(self, example_unitless_da):
         orig = example_unitless_da
         ureg = UnitRegistry(force_ndarray=True)
         result = orig.pint.quantify(ureg.Unit("m"), unit_registry=ureg)
         assert_array_equal(result.data.magnitude, orig.data)
         assert result.data.units == ureg.Unit("m")
+
+    @pytest.mark.parametrize("no_unit_value", conversion.no_unit_values)
+    def test_override_units(self, example_unitless_da, no_unit_value):
+        orig = example_unitless_da
+        result = orig.pint.quantify(no_unit_value, u=no_unit_value)
+
+        with pytest.raises(AttributeError):
+            result.data.units
+        with pytest.raises(AttributeError):
+            result["u"].data.units
 
     def test_error_when_already_units(self, example_quantity_da):
         da = example_quantity_da
@@ -246,6 +263,21 @@ class TestQuantifyDataSet:
         result = orig.pint.quantify({"users": dimensionless})
         assert_array_equal(result["users"].data.magnitude, orig["users"].data)
         assert str(result["users"].data.units) == "dimensionless"
+
+    def test_attach_units_from_str_attr_no_unit(self, example_unitless_ds):
+        orig = example_unitless_ds
+        orig["users"].attrs["units"] = "none"
+        result = orig.pint.quantify({"users": "m"})
+        assert_array_equal(result["users"].data.magnitude, orig["users"].data)
+        assert str(result["users"].data.units) == "meter"
+
+    @pytest.mark.parametrize("no_unit_value", conversion.no_unit_values)
+    def test_override_units(self, example_unitless_ds, no_unit_value):
+        orig = example_unitless_ds
+        result = orig.pint.quantify({"users": no_unit_value})
+        assert (
+            getattr(result["users"].data, "units", "not a quantity") == "not a quantity"
+        )
 
     def test_error_when_already_units(self, example_quantity_ds):
         with raises_regex(ValueError, "already has units"):